[FEE WARNING] "Tiered" fee warning system (#1396)
* move HighFeeWarning to proper file

* proper import paths

* change FEE_SIZE_THRESHOLD to min 10%

* move StyledInfo to own file
