+++ src/custom/pages/Swap/styleds.tsx
@@ -1,0 +1,12 @@
+import styled from 'styled-components'
+import { Info } from 'react-feather'
+
+export const StyledInfo = styled(Info)`
+  opacity: 0.4;
+  color: ${({ theme }) => theme.text1};
+  height: 16px;
+  width: 16px;
+  :hover {
+    opacity: 0.8;
+  }
+`
